Finally actually working
I failed to take two potential cases into account when programming my algorithm class. I have now fixed this. I've also added a randomize function and a function to have a cube follow a pre-determined set of turns. I've also created a simple driver program to test thousands of random cubes. After testing ~20000 (without a failure to solve) I was able to determine that the upper bound on turns is somewhere close to 200 and the average number of turns taken is close to 70.