Proof-of-concept hardware error handler
This file is a proof-of-concept handler that gives us a framework for pausing the bot's execution, resetting the cube, and continuing on where we left off and/or repeating screwed up moves. This mechanism will failure-proof our device at the cost of it no longer being fully automated. At minimum it can act as an emergency stop button or as a way to interact with the bot when it is already running.